+++ packages/create-amplify/templates/basic-auth-data/amplify/data/resource.ts
@@ -1,11 +1,14 @@
 import { defineData } from '@aws-amplify/backend-graphql';
+import { type ClientSchema, a } from '@aws-amplify/amplify-api-next-alpha';
 
-const schema = `
-  type Todo @model @auth(rules: [{ allow: private }]) {
-    id: ID!
-    name: String!
-    description: String
-  }
-`;
+const schema = a.schema({
+  Todo: a.model({
+    id: a.id(),
+    name: a.string(),
+    description: a.string().optional(),
+  }),
+});
+
+export type Schema = ClientSchema<typeof schema>;
 
 export const data = defineData({ schema });