+++ packages/deployed-backend-client/src/deployed_backend_client.test.ts
@@ -27,6 +27,7 @@ import { AwsCredentialIdentityProvider } from '@aws-sdk/types';
 import { DefaultBackendOutputClient } from './backend_output_client.js';
 import { DefaultDeployedBackendClient } from './deployed_backend_client.js';
 import { StackIdentifier } from './index.js';
+import { GetObjectCommand, S3 } from '@aws-sdk/client-s3';
 
 const listStacksMock = {
   NextToken: undefined,
@@ -53,63 +54,6 @@ const listStacksMock = {
       StackName: 'amplify-test-testBranch-api',
       StackStatus: StackStatus.CREATE_FAILED,
       ParentId: 'testStackId',
-    },
-  ],
-};
-
-const describeStacksMock = {
-  Stacks: [
-    {
-      StackName: 'amplify-test-testBranch',
-      StackStatus: StackStatus.CREATE_COMPLETE,
-      StackId: 'testStackId',
-      ParentId: undefined,
-      Outputs: [
-        {
-          OutputKey: 'webClientId',
-          OutputValue: 'webClientIdValue',
-        },
-        {
-          OutputKey: 'awsAppsyncApiEndpoint',
-          OutputValue: 'https://example.com/graphql',
-        },
-        {
-          OutputKey: 'bucketName',
-          OutputValue: 'storageBucketNameValue',
-        },
-        {
-          OutputKey: 'awsAppsyncApiId',
-          OutputValue: 'apiIdValue',
-        },
-        {
-          OutputKey: 'awsAppsyncAuthenticationType',
-          OutputValue: 'AMAZON_COGNITO_USER_POOLS',
-        },
-        {
-          OutputKey: 'authRegion',
-          OutputValue: 'us-east-1',
-        },
-        {
-          OutputKey: 'amplifyApiModelSchemaS3Uri',
-          OutputValue: 's3://schema.graphql',
-        },
-        {
-          OutputKey: 'userPoolId',
-          OutputValue: 'us-east-1_HNkNiDMQF',
-        },
-        {
-          OutputKey: 'identityPoolId',
-          OutputValue: 'us-east-1:identity-pool-id',
-        },
-        {
-          OutputKey: 'storageRegion',
-          OutputValue: 'us-east-1',
-        },
-        {
-          OutputKey: 'awsAppsyncRegion',
-          OutputValue: 'us-east-1',
-        },
-      ],
     },
   ],
 };
@@ -158,6 +102,7 @@ const getOutputMockResponse = {
   [graphqlOutputKey]: {
     payload: {
       awsAppsyncApiEndpoint: 'testAwsAppsyncApiEndpoint',
+      amplifyApiModelSchemaS3Uri: 's3://bucketName/filePath',
     },
   },
 };
@@ -179,6 +124,10 @@ const expectedMetadata = {
     graphqlEndpoint: 'testAwsAppsyncApiEndpoint',
     lastUpdated: undefined,
     status: BackendDeploymentStatus.FAILED,
+    defaultAuthType: undefined,
+    additionalAuthTypes: [],
+    graphqlSchema: 's3://bucketName/filePath schema contents!',
+    conflictResolutionMode: undefined,
   },
 };
 
@@ -186,6 +135,17 @@ void describe('Deployed Backend Client', () => {
   const getOutputMock = mock.fn();
   let deployedBackendClient: DefaultDeployedBackendClient;
   const cfnClientSendMock = mock.fn();
+  const s3ClientSendMock = mock.fn();
+  s3ClientSendMock.mock.mockImplementation((input: GetObjectCommand) => {
+    return {
+      Body: {
+        transformToString: () =>
+          `s3://${input.input.Bucket as string}/${
+            input.input.Key as string
+          } schema contents!`,
+      },
+    };
+  });
 
   beforeEach(() => {
     const mockCredentials: AwsCredentialIdentityProvider = async () => ({
@@ -193,11 +153,14 @@ void describe('Deployed Backend Client', () => {
       secretAccessKey: 'secretAccessKey',
     });
     const mockCfnClient = new CloudFormation();
+    const mockS3Client = new S3();
     getOutputMock.mock.mockImplementation(() => getOutputMockResponse);
     mock.method(mockCfnClient, 'send', cfnClientSendMock);
+    mock.method(mockS3Client, 'send', s3ClientSendMock);
 
     getOutputMock.mock.resetCalls();
     cfnClientSendMock.mock.resetCalls();
+    s3ClientSendMock.mock.resetCalls();
     const mockImplementation = (
       request:
         | ListStacksCommand
@@ -210,7 +173,7 @@ void describe('Deployed Backend Client', () => {
         const matchingStack = listStacksMock.StackSummaries.find((stack) => {
           return stack.StackName === request.input.StackName;
         });
-        const stack = matchingStack ?? describeStacksMock;
+        const stack = matchingStack;
         return {
           Stacks: [stack],
         };
@@ -234,7 +197,8 @@ void describe('Deployed Backend Client', () => {
 
     deployedBackendClient = new DefaultDeployedBackendClient(
       mockCredentials,
-      mockCfnClient
+      mockCfnClient,
+      mockS3Client
     );
   });
 
@@ -302,6 +266,7 @@ void describe('Deployed Backend Client pagination', () => {
   let deployedBackendClient: DefaultDeployedBackendClient;
   const listStacksMockFn = mock.fn();
   const cfnClientSendMock = mock.fn();
+  const s3ClientSendMock = mock.fn();
   const getOutputMock = mock.fn();
   const returnedSandboxes = [
     {
@@ -343,7 +308,9 @@ void describe('Deployed Backend Client pagination', () => {
       ) => DefaultBackendOutputClient;
 
     const mockCfnClient = new CloudFormation();
+    const mockS3Client = new S3();
     mock.method(mockCfnClient, 'send', cfnClientSendMock);
+    mock.method(mockS3Client, 'send', s3ClientSendMock);
     listStacksMockFn.mock.resetCalls();
     listStacksMockFn.mock.mockImplementation(() => {
       return listStacksMock;
@@ -359,7 +326,7 @@ void describe('Deployed Backend Client pagination', () => {
         const matchingStack = listStacksMock.StackSummaries.find((stack) => {
           return stack.StackName === request.input.StackName;
         });
-        const stack = matchingStack ?? describeStacksMock;
+        const stack = matchingStack;
         return {
           Stacks: [stack],
         };
@@ -371,7 +338,8 @@ void describe('Deployed Backend Client pagination', () => {
 
     deployedBackendClient = new DefaultDeployedBackendClient(
       mockCredentials,
-      mockCfnClient
+      mockCfnClient,
+      mockS3Client
     );
   });