+++ packages/backend-function/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @aws-amplify/backend-function
 
+## 0.1.1-alpha.5
+
+### Patch Changes
+
+- 36d93e46: add license to package.json
+- Updated dependencies [36d93e46]
+  - @aws-amplify/backend-output-storage@0.1.1-alpha.2
+  - @aws-amplify/function-construct-alpha@0.1.1-alpha.3
+  - @aws-amplify/plugin-types@0.2.0-alpha.7
+
 ## 0.1.1-alpha.4
 
 ### Patch Changes