+++ packages/backend-output-storage/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @aws-amplify/backend-output-storage
 
+## 0.1.1-alpha.2
+
+### Patch Changes
+
+- 36d93e46: add license to package.json
+- Updated dependencies [36d93e46]
+  - @aws-amplify/backend-output-schemas@0.2.0-alpha.5
+
 ## 0.1.1-alpha.1
 
 ### Patch Changes