+++ packages/form-generator/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @aws-amplify/form-generator
 
+## 0.2.0-alpha.3
+
+### Patch Changes
+
+- 36d93e46: add license to package.json
+
 ## 0.2.0-alpha.2
 
 ### Minor Changes